Replaced compilation_database handler from arduino-cli
